ambugousエラーに関する処理を修正
