+++ Makefile
@@ -1,7 +1,7 @@
-#CC						=	gcc
-CC						=	clang
+CC						=	gcc
+#CC						=	clang
 CFLAGS					=	-Wall -Wextra -Werror
-CFLAGS					+=	-fsanitize=leak
+#CFLAGS					+=	-fsanitize=leak
 #CFLAGS					+=	-fsanitize=address
 INCLUDE					=	-I includes -I ${shell brew --prefix readline}/include
 READLINE_LIB			=	-l readline -L ${shell brew --prefix readline}/lib
@@ -60,7 +60,9 @@ UTILS_SRCS				=	srcs/utils/is_func.c\
 							srcs/utils/call_builtin.c \
 							srcs/utils/put_error_exit.c\
 							srcs/utils/exit_status.c\
-							srcs/utils/sigint_handler.c
+							srcs/utils/sigint_handler.c\
+							srcs/utils/split_lst.c
+
 
 UTILS_OBJS				=	$(addprefix $(OBJS_DIR)/, $(UTILS_SRCS:.c=.o))
 #####################################
