+++ srcs/expander/expander.c
@@ -1,6 +1,6 @@
 #include "minishell.h"
 
-static t_list	*expand_cmd_info_element(t_list *list, char **my_env)
+t_list	*expand_cmd_info_element(t_list *list, char **my_env)
 {
 	t_list	*new_lst;
 	t_list	*current;
@@ -12,19 +12,13 @@ static t_list	*expand_cmd_info_element(t_list *list, char **my_env)
 	while (current != NULL)
 	{
 		token = current->content;
-		if (!is_heredoc(token))
+		ms_lstadd_back(&new_lst, expand(token, FALSE, my_env));
+		tmp = new_lst;
+		while (tmp != NULL)
 		{
-			ms_lstadd_back(&new_lst, expand(token, FALSE, my_env));
-			tmp = new_lst;
-			while (tmp != NULL)
-			{
-				if(!is_heredoc(tmp->content))
-					tmp->content = remove_quotation(tmp->content);
-				tmp = tmp->next;
-			}
+			tmp->content = remove_quotation(tmp->content);
+			tmp = tmp->next;
 		}
-		else
-			ms_lstadd_back(&new_lst, ms_lstnew(ms_strdup(token)));
 		current = current->next;
 	}
 	ms_lstclear(&list, free);
